Version 1.5.0
In this new version, ACL capabilities are started to be supported in SilverChat
in order to restrict some of the functionalities. For instance, only an ACE is
added and it is about the capability to create a groupchat (and hence to invite
a buddy to join it). By default, no restriction.