Version 1.5.1
When a groupchat is deleted by its owner, all of the members of the group
receive a presence stanza that is taken in charge explicitly by SilverChat
in order to unbookmark the deleted room. But, the code in jsxc listening the
incoming presence stanza is executed after the SivlerChat's one and that causes
in this situation an error because the room doesn't exist anymore and the jsxc
code expects to be here. As a workaround of this issue, the unbookmarking in
SilverChat is executed within a setTimeout instruction after 1s; this leaves
the time for the JSXC handler to be completed.

When a user delete a groupchat, all the connected members of this room are
notified about that and then the room is automatically removed from their own
bookmark. But the non-connected members aren't notified. So, in order to remove
the deleted room from their bookmark at their connection, the JSXC function
jsxc.xmpp.bookmarks.loadFromRemote has been overridden to load all the
available rooms from the remote MUC service and then check each of the
bookmarked room is included in this list otherwise the room is unbookmarked.

If a user invite one of his contact that is disconnected to a group chat but
then delete the group chat, then the invitation is well received by the contact
at his connection and then automatically SilverChat accepts the invitation by
joining the deleted room: this bug is now fixed by here also listing the
available rooms in order to decide to join or not the room.